+++ doc/changelog.md
@@ -13,7 +13,7 @@ provided.
 ## {{ next-release }}
 
 - Add
-    - [`module/bench`]
+    - [`module/lab.bench`]
     - [`module/maestro.idiom`]
 - Impacted
     - [`module/string`]
@@ -48,17 +48,17 @@ Inception.
 
 
 [`module/bb.help`]:                  ../module/bb.help/doc/changelog.md
-[`module/bench`]:                    ../module/bench/doc/changelog.md
 [`module/classpath`]:                ../module/classpath/doc/changelog.md
 [`module/deps.edn`]:                 ../module/deps.edn/doc/changelog.md
 [`module/edn`]:                      ../module/edn/doc/changelog.md
 [`module/git`]:                      ../module/git/doc/changelog.md
-[`module/maestro`]:                  ../module/maestro/doc/changelog.md
+[`module/lab.bench`]:                ../module/lab.bench/doc/changelog.md
 [`module/maestro.idiom`]:            ../module/maestro.idiom/doc/changelog.md
 [`module/maestro.plugin.build`]:     ../module/maestro.plugin.build/doc/changelog.md
 [`module/maestro.plugin.clj-kondo`]: ../module/maestro.plugin.clj-kondo/doc/changelog.md
 [`module/maestro.plugin.kaocha`]:    ../module/maestro.plugin.kaocha/doc/changelog.md
 [`module/maestro.plugin.quickdoc`]:  ../module/maestro.plugin.quickdoc/doc/changelog.md
+[`module/maestro`]:                  ../module/maestro/doc/changelog.md
 [`module/namespace`]:                ../module/namespace/doc/changelog.md
 [`module/process`]:                  ../module/process/doc/changelog.md
 [`module/string`]:                   ../module/string/doc/changelog.md